Now writing HUDS to video.
This raises some important questions that have to be addressed later.
What happens when an error occurs?
Ideally I should write out the original image, and the HUDS display should
announce that there is an error and display the original image and the illumination file but should also announce that there was an error.

That will be in the next commit.