Part way through setting up a new system whereby main.cpp, when passed a compiler directive, -Dsimulate will now run and compile in a hardware-independant fashion.
This means it will no longer require the Camera or DLP libraries to be around. This means that someone could get everything working with only the git repository and OpenCV.

The trick is in the use of DontTalk2DLP and DontTalk2Camera which are substituted for Talk2DLP.c and Talk2Camera.c in the makefile.

As it stands, this has not been tested. But the idea is that now main.cpp will be compiled inot ClosedLoop.exe and SegmentFrame.exe

It will be basically the same except SegmentFrame.exe will read from video.