Removed Critical Intermittant Failure with FrameGrabber frame shifting (II).
The FrameGrabber seems to be inconsistantly applying the command to shift the frame in y by 127 pixels. This leads to serious problems. So we removed the line that told the camera to shift the region of interest by 127 pixels. This is somewhat unfortunate because it means that when we crop the camera's CCD field from 1280x1024 to 1024x768 we are actually only using the upper left hand region of the camera (or some such oddity.) But for now this will have to do. Later one we shall need to perform more investigation to find out why the FrameGrabberSetRegionOfInterest function fails intermittantly.