Tried tweak so that one thread won't crash when the other stops.
I think what happens is the main thread stops and then immediate clears out the memory for the experiment variable. The display thread doesn't realize the main thread had stopped and it tries to access a variable that doesn't exist and crashes.

I've tried to fix this by adding a dealy between the first thread announcing that its done and when it goes and frees up the memory.