POTENTIALLY RESOLVED CRASH BUG
I did a number of things to potentially resolve the big that was going on. There were two main components to this bug. The first was that when acquisition would fail it would hang the process for a long time. The second was that when it recovered it shifted the frame up by a few 100 pixels rendering the rest of the experiment useless.

I beleive I have fixed both of these problems. To fix the frame shifted, I call CiAqCommand() with the paramater CiConReset and tell the framegrabber board to reset. When an acquisition error occurs I also now have in place the infrastructure to handle that error by immediaetly requesting another frame.

I also fixed the long duration of hangs by using the CiBrdAqTimeoutSet command when setting up the framegrabber and providing a timeout duration of 4ms. (I might make this even shorter, but for now 4ms seems quite acceptable.)

More testing is required, but I believe this should resolve the problem.

-Andy