Fix potentially critical bug
On some systems, when using an illumination protocol, errant patterns may be being sent to the DLP.

These patterns occur because the cvCreateImage function does not seem to always initialize images to zero. So here we have added cvSetZero to set the temp image to zero and this has fixed the problem.